Embedding/requirements.txt
google-genai>=1.0.0
pinecone>=5.0.0
openai>=1.0.0
python-dotenv>=1.0.0
Pillow>=10.0.0
moviepy>=2.0.0
flask>=3.0.0
